vp9: fix buffer overflow in vp9_get_token_cost
Clamp the max_index for the vp9_cat6_high#_cost[] tables.

Also clamp the see calculation in sum_squarss_visible(),
as the test below causes overflow in that function.
Similar to the change in encodeframe.c

Enable the test that triggers the issue.

Bug: 488585490

+++ vp9/encoder/vp9_encodeframe.c
@@ -446,7 +446,7 @@ static void fill_variance(uint32_t s2, int32_t s, int c, Var *v) {
 static void get_variance(Var *v) {
   v->variance =
       (int)(256 * (v->sum_square_error -
-                   (uint32_t)(((int64_t)v->sum_error * v->sum_error) >>
+                   (uint64_t)(((int64_t)v->sum_error * v->sum_error) >>
                               v->log2_count)) >>
             v->log2_count);
 }
